requirements.txt
fastapi==0.121.0
pandas==2.3.3
uvicorn==0.38.0
psycopg2-binary==2.9.11
boto3==1.34.34
python-dotenv==1.0.0